restrucutre files and folder to make it easier to other to update
